==> Head/2Scripts/01.FulltreeCodonsAnalysis.R <==
# Start fresh by cleaning your environment

# 0. First check dataset for disrepancies (на вшивость) 

# ANALYSIS

# 1. Simple summary descriptive statistics
# How many different positions in codon

# 2. Where along the genome
# histogram on genome + 2000 genomes article
# Which genes
# are they correlated (Spearman rank and scatterplot )

# 2. Loss of function or gain of function?
# How many stop codons or not?
# How many to stop or from stop

# 3. Which aa to which
# in which biochem group using matrices Grantham distance
# How many ambiguous
